+++ main.go
@@ -22,6 +22,7 @@ import (
 	"github.com/matthieukhl/rgvr/cmd"
 	_ "github.com/matthieukhl/rgvr/cmd/auth"
 	_ "github.com/matthieukhl/rgvr/cmd/groups"
+	_ "github.com/matthieukhl/rgvr/cmd/numbers"
 	_ "github.com/matthieukhl/rgvr/cmd/teams"
 	_ "github.com/matthieukhl/rgvr/cmd/users"
 )