+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `numbers list` command to retrieve the team's phone numbers.
+
+### Fixed
+
+- Fixed an issue in which the `groups` command returned a null `users` key.
+
 ## [0.4.0-beta] - 2026-08-02
 
 ### Added